+++ changelog.md
@@ -15,6 +15,7 @@
 * Reduce calls to the Aura API during GdsSessions::get_or_create.
 * Improve error message when a query is interrupted by a signal (SIGINT or SIGTERM).
 * Improve error message if session is expired.
+* Improve robustness of Arrow client against connection errors such as `FlightUnavailableError` and `FlightTimedOutError`.
 
 
 ## Other changes